Factored out ResultIntentHelper
